convert-data: Split log into debci and autopkgtest parts
Try to find a "Started" log section and place that into /autopkgtest/, and keep
all the other sections in the debci log. This mirrors how we create new logs
now.

The format changed a bit over time (two different separators, different
headings, etc.). Fortunately the "Started" record never changed, so we take
that as the start of the adt-run log until the next record.

A few recent logs won't have a "Started" section due to #747041. These logs are
pretty much broken and there's not a lot we can do to salvage them, so just
keep them completely as debci log and don't create an /autopkgtest log.