Add tests for debci-batch
These coverskipping/re-running of tests under appropriate conditions and
showing the cause for re-run in the log.

Add a status_dir_for_package() to test_helper.sh that works like the "real" one
in environment.sh, to avoid putting the knowledge of the data/ dir structure in
too many places.