status: fix handling of missing status files
When the very first package in the list didn't have a status file yet,
the generated JSON file was invalid.

This issue revealed totally by coincidence when the output of
debci-list-packages was changed to be sorted in the presence of a
whitelist.

We probably need to rewrite debci-status in Ruby soon.