debci-setup/debci-test: avoid concurrent execution
We do not want tests running while the (base) testbed is being updated,
so debci-setup gets an exclusive lock on the testbed.

Multiple tests running concurrently should be fine, hence debci-test
gets a shared lock. For this to work correctly the backend in use has a
notion of a base testbed, and that each test run spawns a temporary
testbed based on that base one. This is the case for all the currently
supported backends.