Sync test results through the AMQP server
Instead of relying on an external filesystem sync to bring data from
multiple workers together, just use a results queue in the AMQP server
for that.

Results are transmitted as a gzip-compressed tarball, which is extracted
by the master server (the one that runs the new debci-collector process)
upon reception.

For now, we assume that all workers need to be trusted, but in the
future we can authenticated the test results using a GPG signature or
something similar.

I was initially wary of the message size capacity of RabbitMQ, but the
largest result directory on ci.debian.net (glibc) compresses down to
~2MB, so we are fine.