Allow URL base config items to refer to suite
Replace "{SUITE}" in site_url and artifacts_url_base config values with the
current suite when rendering pages. This cannot be done in debci-config
already, as generate-html applies to all pages and isn't invoked per suite and
architecture.

This is useful to use one Swift container per suite, which makes it much easier
to clean up obsolete logs.