debci-generate-index: Fix unset extra_fields parameter
When running foreman, the collector reported the following error:

bin/debci-generate-index: extra_fields: parameter not set

Upon further inspection, extra_fields was only set when there was
an $artifacts_url_base. Otherwise, it was not set and execution
stopped when debci-generate-index generated the ${status_file}.

This prevented updates to the web interface.