Debci::Repository#tmpfail_packages: Ensure unique packages
Commit 944671b was thought to have fixed the duplicate package
problem that was experienced in the status alert view on
autopkgtest.ubuntu.com/status/alerts.

However, the problem was not fixed and packages are still reported
N times, where N is the number of unique suites a package is
temporarily failing.

This commit fixes the problem by ensuring uniqueness in the package
data that is returned by Debci::Repository#tmpfail_packages.