services: restart on failure
Usually the only case where amqp-consume will exit with a failure is if
it can't connect to the AMQP server, what may happen when the network is
not up yet, or when the AMQP server restarts.

Also, specify `After=rabbitmq-server.service network-online.target` to
wait for rabbitmq-server it it's installed locally, and for the network
to be up.