debci-batch: don't enqueue packages that are already in the queue
This is valid only in debci-batch; if debci-enqueue is called directly,
it will always enqueue. The goal is to be able to run debci-batch as
frequently as possible, and still not have packages being enqueued a
second time before they got a chance to run since the last time a test
run was requested for it.