debci batch: optimize the test for the need of test requests
Turns out that testing whether a package has already been queued is a
lot faster than checking its list of transitive dependencies, so if we
test for it being already scheduled _before_ testing the dependencies,
we can short circuit the test and save a lot of time. This makes it
possible to run debci batch very frequently.