debci-list-dependencies: exclude packages from the base system
When we started calculating dependencies from the chdist instead of from
an actual chroot, we started listing as dependencies *all packages* that
would be installed in a _empty_ system to install the given package,
which would of source include packages from the base system.

We now reconstruct the base system by listing all required|important
packages, plus dpkg-dev, with their dependencies. This matches exactly
the packages that are installed in a fresh debci testbed.