Fix default mirror for setup-chdist
Revert commit 72673a36 which sets a default mirror in environment.sh, as this
short-circuits the mirror detection in setup-chdist. We could do the latter in
environment.sh, but as this is relatively expensive, just leave the default
empty. Set the default in schroot/create-testbed instead.

For consistency, change the default mirror of setup-chdist to
http://http.debian.net/debian.