debci-expire: control data retention
Since we use the plain filesystem for storage, it comes a point where
there are not enough inodes in the partition to store all data.
debci-expire will remove the following files when they are older than a
configurable amount of days (180 by default):

- autopkgtest artifacts and logs
- debci logs

Pending: remove the json result files for each test run.