Add support for migration tests
A migration is a test run test where you want to check if package X can
be migrated from suite A (e.g. Debian unstable) to suite B (e.g. Debian
testing). In this case, you run the tests on a base test bed from suite
B, but add APT sources for suite A and pin the packages you want to
migrate to be obtained from there. This can be used to test either new
versions of package X itself, or packages that depend on it.

Beying unstable → testing migration, this can also be used to test
security updates, backports, etc.

debci-test, and debci-enqueue can now take two extra options:

--trigger: specifies the reason why this test is being performed. For
debci, this is just a string and has no special meaning. This string is
stored together with the test artifacts, and will also eventually be
displayed in JSON data, and in the UI.

--pin-packages=RELEASE=pkg,src:pkg. Adds APT sources for RELEASE, and
pins the specified packages to be obtained from RELEASE. `src:pkg`
expand to all binary packages from source package `pkg`. This requires
autopkgtest >= 5.1.