debci-generate-index: make sure to always run
Skipping a run when the lock is obtained by another process cause
results to accumulate in the incoming directory when there is heavy
traffic of results arriving. If we are unlucky, several subsequent runs
of debci-generate-index will be skipped because there was some test
result being received exactly at that time.

This will cause the results queue to grow a little bit while
debci-generate-index is running.