Escape triggers when enqueuing test jobs
Trigger with spaces (or worse, with shell hackery) were being passed to
workers as is, and then used to compose a shell command line.

Instead, when enqueing the job we CGI-escape them so only ascii
characters remain untouched. Then, when reading the trigger back from
the data file, we unescape it.

Closes: #902337