Move data from JSON files to database
This helps us reduce the inode usage on the debci master host. Instead
of keeping around 1 JSON file for each test run, move all corresponding
test data to the database.

A copy of what would be that JSON data file is still kept around as
"latest.json", which used to be a symlink to the newest test job JSON
file.

"history.json" is also kept for backwards compatibility, but it's now
produced from the database instead of by reading all of the JSON files.