lxc/create-testbed: avoid cleaning up valid containers
Sometimes, the creation of a new testbed fails for various reasons (some
essential package) breaks debootstrap, or something fails to download.
Those failures are already handled by autopkgtest-build-lxc: the old
container is only replaced by the new one in case of success.  In that
case, it's OK for debci to just keep using the older container since it
will be upgraded before running any tests anyway.

The only case where we indeed want to remove containers is when the
customizations needed by debci are not successfully applied.