config: fix setting debci_data_retention_days
The defaults are erroneously defined *after* configuration files are
loaded, so forgetting to check an existing value before settings causes
settings to not be configurable. In the long run we want to rewrite that
part to set the default before loading the config files, but this needs
to be fixed right now.