test_worker: exclude existing processes from check for leftover processes
This is a long-standing problem, which caused the tests to fail when run
in parallel with an existing set of running debci daemons (e.g. when
your have the daemons running in one terminal, and runs the test suite
in another).

This has just manifested when running the test suite under autopkgtest,
since we now need debci-collector installed and that starts the
collector daemon in the background.