debci-test: add buildd suite for migration tests
This prevents the situation where a migration test gets the source
package from unstable, but the corresponding binaries are not yet
available, so the test ends up running against the binaries from
testing

britney only schedules tests when the mirror it uses already has the
corresponding binaries, but the mirror used by the workers can still be
outdated. By adding the buildd suites from incoming, we ensure that
those binaries will be found.