Make test suite pass on non-amd64
All the changes are basically replacing hardcoded "amd64" with an
expression that returns the architecture of the current system. There
are still other hardcoded "amd64" references out there, but those are
not a problem at the moment as the context in which they are used does
not depend on the current host architecture.