Debci::Package#backend: make test independent of local configuration
Debci::Config.backend_list is resolved at startup, so if your local
configuration does not specify `lxc` as a backend, then the test fails.
This can be reproduced just adding an explicit test for the validation
of `qemu` in the test - this would pass if I had `qemu` in my local
`backend_list`.

The solution is making the inclusion list be calculated at runtime with
a lambda. This makes the mocking done in the test be actually used.