Debci::Job: add index for field package_id
There used to be an index on jobs.package back in the day. In
e1723e9b3589e677de5454fc8cff2495bcd8f35e, that was changed to use a
proper foreign key, but that new field never got an index for it.

This greatly improves performance when listing jobs per package, as well
as when receiving new test results.