Debci::Job index requestor_id
Same thing as for package_id: we used to have a `requestor` column,
which *was* indexed. It was then migrated to a normalized foreign key
`requestor_id`, but we missed indexing it. This will vastly reduce query
times for jobs by their requestor, e.g. API calls for submitted jobs and
the self service UI.